requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.0
PyJWT>=2.8
mysqlclient>=2.2
fastapi>=0.105
uvicorn>=0.23
pydantic>=2.6
pytest>=7.4
pytest-django>=4.7
requests>=2.31
python-dotenv>=1.0
